Safari-611.1.10/LayoutTests/svg/custom/clone-node.html

<!DOCTYPE html>
<script src="../../resources/js-test-pre.js"></script>
<script>

description('Tests cloneNode for SVGDocument.');

function className(object) {
    return Object.prototype.toString.call(object).slice(8, -1);
}

var svgDoc = document.implementation.createDocument('http://www.w3.org/2000/svg', 'svg', null);

shouldBe('svgDoc.cloneNode(false).__proto__', 'XMLDocument.prototype');
shouldBeEqualToString('className(svgDoc.cloneNode(false))', 'XMLDocument');
shouldBeEqualToString('svgDoc.cloneNode(true).compatMode', 'CSS1Compat');

</script>
<script src="../../resources/js-test-post.js"></script>
