safari-600.7-branch/LayoutTests/svg/custom/clone-node-expected.txt

Tests cloneNode for SVGDocument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svgDoc.cloneNode(false).__proto__ is SVGDocument.prototype
PASS className(svgDoc.cloneNode(false)) is "SVGDocument"
PASS svgDoc.cloneNode(true).compatMode is "CSS1Compat"
PASS successfullyParsed is true

TEST COMPLETE
